Mantra.Oracle

Description

Oracle for general data.

Oracle

data OracleScript

Type for the script.

oracleInstance

Arguments

:: Oracle
The oracle.

-> TypedValidator OracleScript
The instance.

Compute the instance for an oracle.

oracleValidator

Arguments

:: Oracle
The oracle.

-> Validator
The validator.

Compute the validator for an oracle.

oracleAddress

Arguments

:: NetworkId
The network identifier.

-> Oracle
The oracle.

-> AddressAny
The script address.

Compute the address for an oracle.

plutusOracle

Arguments

:: Oracle
The oracle.

-> PlutusScript PlutusScriptV1
The Plutus script.

Serialise the oracle as a Plutus script.

exportOracle

Arguments

:: FilePath
The filename for writing the validator bytes.

-> NetworkId
The network identifier.

-> Oracle
The oracle.

-> IO AddressAny
Action writing the validator and returning its address.

Export the validator for an oracle and compute its address.

Access

fetchDatum

Arguments

:: TxOut
The transaction output.

-> (DatumHash -> Maybe Datum)
Function for looking up the datum, given its hash.

-> Maybe BuiltinData
The datum, if it was found.

Retrieve the oracle's datum from a transaction output.

Orphan instances

FromData Action

Instance details

Methods

fromBuiltinData :: BuiltinData -> Maybe Action

ToData Action

Instance details

Methods

toBuiltinData :: Action -> BuiltinData

UnsafeFromData Action

Instance details

Methods

unsafeFromBuiltinData :: BuiltinData -> Action

Lift DefaultUni Oracle

Instance details

Methods

lift :: Oracle -> RTCompile DefaultUni fun (Term TyName Name DefaultUni fun ())

Typeable DefaultUni Oracle

Instance details

Methods

typeRep :: Proxy Oracle -> RTCompile DefaultUni fun (Type TyName DefaultUni ())