Mantra.Oracle.Types

Description

Types describing the general-purpose datum oracle.

Types

data Oracle

An oracle controlled by one token, holding another token, and requiring a fee for its use.

Constructors

Oracle

Fields

controlToken :: !AssetClass

The token needed for writing or deleting the oracle.

datumToken :: !AssetClass

The token with which the oracle datum is always associated.

requiredFee :: !Value

The minimum fee required to read the oracle on-chain.

data Parameters

Parameters defining the oracle.

Constructors

Parameters

Fields

controlParameter :: AssetClass

The token needed for writing or deleting the oracle.

datumParameter :: AssetClass

The token with which the oracle datum is always associated.

feeToken :: AssetClass

The token in which the fee must be paid for reading the oracle.

feeAmount :: Integer

The amount of the fee token needed for reading the oracle, if any.

lovelaceAmount :: Integer

The amount of Lovelace needed for reading the oracle, if any.

data Action

Redeemers for the oracle.

Constructors

Delete
Delete (close) the oracle.

Read
Read the oracle's datum.

Write
Set or update the oracle's datum.

Constructors

makeOracle

Arguments

:: Parameters
Parameters describing the oracle.

-> Oracle
The oracle.

Construct an oracle.